backend/requirements.txt
# Core
Flask==3.1.0
flask-cors==5.0.1
Werkzeug==3.1.3

# Data Processing
pandas==2.2.3
numpy==1.26.0  # Downgraded for compatibility

# Machine Learning
scikit-learn==1.2.2  # Must match training version
joblib==1.2.0  # Compatible with scikit-learn 1.2.2
scipy==1.10.1  # Compatible version

# Utilities
python-dotenv==1.1.0